Mayor Update! V1.63b - Added AMBE-addon modules - Added EEProm / Config-Storage support - split main file 'rptr_main.c' into 3 modules:   'rptr_main.c': init + main-loop   'config.c'  : load/save config blocks from/to EEProm, configure functions   'stdmodem.c' : "behavior" functions, communication with PC

+++ trunk/src/opendv/dv_dstar.h
@@ -97,7 +97,7 @@ typedef struct __attribute__((__packed__)) {
 typedef char tSlowDataFrame[60];
 
 
-void	dstar_buildheader(unsigned long *wrbuffer, const tds_header *header);
+void	dstar_buildheader(unsigned int *wrbuffer, const tds_header *header);
 
 // Creates a 60byte slow-data stream filled with the radio header (used 4 idle data):
 // function is obsolete
